+++ src/main/java/org/example/dto/PostDTO.java
@@ -1,26 +1,26 @@
 package org.example.dto;
 
 public class PostDTO {
-    private Integer id;
+    private Long id;
     private String content;
-    private Integer authorId;
+    private Long authorId;
     private String authorName;
 
-    public PostDTO(Integer id, String content, Integer authorId, String authorName) {
+    public PostDTO(Long id, String content, Long authorId, String authorName) {
         this.id = id;
         this.content = content;
         this.authorId = authorId;
         this.authorName = authorName;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getContent() { return content; }
     public void setContent(String content) { this.content = content; }
 
-    public Integer getAuthorId() { return authorId; }
-    public void setAuthorId(Integer authorId) { this.authorId = authorId; }
+    public Long getAuthorId() { return authorId; }
+    public void setAuthorId(Long authorId) { this.authorId = authorId; }
 
     public String getAuthorName() { return authorName; }
     public void setAuthorName(String authorName) { this.authorName = authorName; }
