+++ src/main/java/org/example/dto/UserDTO.java
@@ -1,18 +1,18 @@
 package org.example.dto;
 
 public class UserDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String login;
 
-    public UserDTO(Integer id, String name, String login) {
+    public UserDTO(Long id, String name, String login) {
         this.id = id;
         this.name = name;
         this.login = login;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
